+++ sdk/healthdataaiservices/Azure.Health.Deidentification/CHANGELOG.md
@@ -2,5 +2,8 @@
 
 ## 1.0.0-beta.1 (2024-08-15)
 
+- Azure Health Deidentification client library
+
 ### Features Added
-- Initial Creation
+
+- Azure Health Deidentification client library